Refactor: util 폴더 정리
