--- src/utils/isNumber.ts
@@ -1,7 +1,0 @@
-export const isNumber = (value: string | number) => {
-  if (typeof value === 'number') {
-    return /^[0-9]+$/.test(value.toString());
-  } else {
-    return /^[0-9]+$/.test(value);
-  }
-};
